+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.1.1
 - Added "Always on top" option in settings
+- Fixed api control action bug
 
 ## 1.1.0
 - Implements new "api_control" mode to utilize the API instead of the mod. This uses the same control name but sends the value using the HTTP API (may require the API key to be configured).